adjusted requested PR changes
