+++ frontend/app/components/layout/AppHeader.vue
@@ -1,4 +1,6 @@
 <script setup lang="ts">
+import HelpdeskFormModal from '~/components/helpdesk/HelpdeskFormModal.vue'
+
 const authentication = useAuthentication()
 const llmStore = useLLM()
 const router = useRouter()
@@ -34,7 +36,7 @@ const dropdownItems = computed(() => [
     </template>
 
     <template #right>
-      <NeedHelpCard />
+      <HelpdeskFormModal />
       <UColorModeSwitch />
 
       <UButton
